Fix ninja dependency issue.
The real fix is the nogncheck annotation and the comment explaining why it's
there.

The other changes are to ensure that we always set
TIME_SYNCHRONIZATION_CLUSTER_INCLUDED in chip_data_model.gni and check for it
using #if, not #ifdef.  That should catch issues with build systems that are not
setting it at all, as long as -Wundef is being used there, and lead to that
being fixed.